manipula_poligono/requirements.txt
PyOpenGL
PyOpenGL_accelerate
pytest
black
ruff
